bugs exercicios

+++ src/persistencia/ExercicioDAO.java
@@ -227,15 +227,16 @@ public class ExercicioDAO {
 
 		File dir3 = new File("banco/desempenhos/" + user.getMatricula()	+ "/acertos/" + exercicio.getCode() + ".csv");
 		FileWriter writer;
-		
-			try {
-				writer = new FileWriter(dir3);
-				writer.write(exercicio.getCode());
-				writer.write(";");
-				writer.flush();
-				writer.close();
-			} catch (IOException e) {
-				e.printStackTrace();
+			if(!dir3.exists()){
+				try {
+					writer = new FileWriter(dir3);
+					writer.write(exercicio.getCode());
+					writer.write(";");
+					writer.flush();
+					writer.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
 			}
 		
 	}
@@ -252,6 +253,7 @@ public class ExercicioDAO {
 
 		File dir3 = new File("banco/desempenhos/" + user.getMatricula()	+ "/erros/" + exercicio.getCode() + ".csv");
 		FileWriter writer;
+		if(!dir3.exists()){
 			try {
 				writer = new FileWriter(dir3);
 				writer.write(exercicio.getCode());
@@ -261,6 +263,7 @@ public class ExercicioDAO {
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
+		}
 		
 	}
 	public int getAcertos(Usuario user) {
